requirements.txt
flask>=3.0.0
requests>=2.31.0
openai>=1.30.0
python-dotenv>=1.0.0
gunicorn>=22.0.0
gspread>=6.0.0
google-auth>=2.29.0
